# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.UpdatePasswordInput:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        minLength: 6
        type: string
    required:
    - new_password
    - old_password
    type: object
  models.UpdateProfileInput:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      username:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - username
    type: object
...
host: localhost:3001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API backend for digital parenting application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Digital Parenting App BE API
  version: "1.0"
paths:
...
  /user/password:
    patch:
      consumes:
      - application/json
      description: Updates the current user's password.
      parameters:
      - description: Password Update Details
        in: body
        name: update_password
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Validation failed or invalid request body
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Invalid old password
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error during password update
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update My Password
      tags:
      - User - Profile Management
  /user/profile:
    get:
      description: Get the profile for the current user.
      produces:
      - application/json
      responses:
        "200":
          description: Profile data for current user
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Validation failed or invalid request body
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Failed to identify user
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error during profile retrieval
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get my profile
      tags:
      - User - Profile Management
    patch:
      consumes:
      - application/json
      description: Update the profile for the current user.
      parameters:
      - description: Profile update information
        in: body
        name: update_profile_input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Validation failed or invalid request body
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Failed to identify user
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error during profile update
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update my profile
      tags:
      - User - Profile Management
securityDefinitions:
  ApiKeyAuth:
    description: '"Type ''Bearer YOUR_JWT_TOKEN'' into the value field."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"